Fix <cmath> include and replace it with <math.h> for Arduino and ESP compatibility

+++ include/cobalt/math/algebra/complex/complex.hpp
@@ -56,7 +56,7 @@ struct Complex {
          *  @param r The norm/magnitude of the complex number
          *  @param theta The argument/angle of the complex number
          */
-        static Complex polar(float r, float theta) noexcept { return Complex(r*static_cast<float>(std::cos(theta)), r*static_cast<float>(std::sin(theta))); }
+        static Complex polar(float r, float theta) noexcept { return Complex(r*cosf(theta), r*sinf(theta)); }
 
 
         // ---------------- Accessors ----------------
